# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * TBD Recommendations API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  name: string;
  createdAt: string;
}

export interface UserInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  name: string;
}

export type SuggestionCategory = typeof SuggestionCategory[keyof typeof SuggestionCategory];


export const SuggestionCategory = {
  book: 'book',
  movie: 'movie',
  tv: 'tv',
} as const;

export interface Suggestion {
  id: number;
  userId: number;
  userName: string;
  category: SuggestionCategory;
  title: string;
  description: string;
  commentCount: number;
  createdAt: string;
}

export type SuggestionInputCategory = typeof SuggestionInputCategory[keyof typeof SuggestionInputCategory];


export const SuggestionInputCategory = {
  book: 'book',
  movie: 'movie',
  tv: 'tv',
} as const;

export interface SuggestionInput {
  userId: number;
  category: SuggestionInputCategory;
  /**
     * @minLength 1
     * @maxLength 200
     */
  title: string;
  /** @minLength 1 */
  description: string;
}

export type SuggestionDetailCategory = typeof SuggestionDetailCategory[keyof typeof SuggestionDetailCategory];


export const SuggestionDetailCategory = {
  book: 'book',
  movie: 'movie',
  tv: 'tv',
} as const;

export interface Comment {
  id: number;
  suggestionId: number;
  userId: number;
  userName: string;
  content: string;
  createdAt: string;
}

export interface SuggestionDetail {
  id: number;
  userId: number;
  userName: string;
  category: SuggestionDetailCategory;
  title: string;
  description: string;
  commentCount: number;
  createdAt: string;
  comments: Comment[];
}

export interface CommentInput {
  suggestionId: number;
  userId: number;
  /** @minLength 1 */
  content: string;
}

export interface Summary {
  total: number;
  books: number;
  movies: number;
  tv: number;
}

export interface ErrorResponse {
  error: string;
}

export type ListSuggestionsParams = {
category?: ListSuggestionsCategory;
};

export type ListSuggestionsCategory = typeof ListSuggestionsCategory[keyof typeof ListSuggestionsCategory];


export const ListSuggestionsCategory = {
  book: 'book',
  movie: 'movie',
  tv: 'tv',
} as const;