# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * TBD Recommendations API
 * OpenAPI spec version: 0.1.0
 */

export interface Comment {
  id: number;
  suggestionId: number;
  userId: number;
  userName: string;
  content: string;
  createdAt: Date;
}