# lib/api-zod/src/generated/types/commentInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * TBD Recommendations API
 * OpenAPI spec version: 0.1.0
 */

export interface CommentInput {
  suggestionId: number;
  userId: number;
  /** @minLength 1 */
  content: string;
}