# lib/api-zod/src/generated/types/suggestion.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * TBD Recommendations API
 * OpenAPI spec version: 0.1.0
 */
import type { SuggestionCategory } from './suggestionCategory';

export interface Suggestion {
  id: number;
  userId: number;
  userName: string;
  category: SuggestionCategory;
  title: string;
  description: string;
  commentCount: number;
  createdAt: Date;
}