# lib/api-zod/src/generated/types/suggestionDetail.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * TBD Recommendations API
 * OpenAPI spec version: 0.1.0
 */
import type { Comment } from './comment';
import type { SuggestionDetailCategory } from './suggestionDetailCategory';

export interface SuggestionDetail {
  id: number;
  userId: number;
  userName: string;
  category: SuggestionDetailCategory;
  title: string;
  description: string;
  commentCount: number;
  createdAt: Date;
  comments: Comment[];
}