# lib/api-zod/src/generated/types/suggestionInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * TBD Recommendations API
 * OpenAPI spec version: 0.1.0
 */
import type { SuggestionInputCategory } from './suggestionInputCategory';

export interface SuggestionInput {
  userId: number;
  category: SuggestionInputCategory;
  /**
     * @minLength 1
     * @maxLength 200
     */
  title: string;
  /** @minLength 1 */
  description: string;
}