# lib/api-zod/src/generated/types/summary.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * TBD Recommendations API
 * OpenAPI spec version: 0.1.0
 */

export interface Summary {
  total: number;
  books: number;
  movies: number;
  tv: number;
}