# lib/api-zod/src/generated/types/userInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * TBD Recommendations API
 * OpenAPI spec version: 0.1.0
 */

export interface UserInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  name: string;
}